move job manager to lib
